# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CognIA API
  version: 0.1.0
  description: |
    API backend de la tesis "Aplicativo web con Random Forest para la alerta temprana
    de cinco trastornos psicologicos en ninos de 6 a 11 anos".
    Proyecto academico del programa de Ingenieria de Sistemas y Computacion en la
    Universidad de Cundinamarca (grupo GISTFA).
    Esta API genera alertas de riesgo y no representa un diagnostico clinico.
    Trabaja solo con datos simulados o anonimizados.

    Enfoque de seguridad y buenas practicas:
    - Auth: JWT access token en JSON; refresh token en cookie HttpOnly con path
      /api/auth/refresh.
    - CSRF: double-submit con header X-CSRF-Token y cookie csrf_refresh_token; la
      cookie rota en refresh.
    - MFA: TOTP con secreto cifrado y recovery codes; obligatorio para roles
      privilegiados.
    - Observabilidad: /healthz (liveness), /readyz (DB readiness), /metrics (token
      opcional, por proceso).

    Notas:
    - Si usas cookies, define CORS_ORIGINS explicitos y envia credenciales.
    - /api/predict es experimental y puede cambiar.

    Creditos:
    - Andres Felipe Melo Chaguala (estudiante investigador)
    - Johan Thomas Cristancho Silva (estudiante investigador)
    - Oscar Jobany Gomez Ochoa (director)
    - Universidad de Cundinamarca, grupo GISTFA
servers:
  - url: /
    description: Same-origin (use the host that serves /docs)
  - url: http://127.0.0.1:5000
    description: Local development (IPv4)
...
paths:
...
  /api/v1/questionnaires/active:
    get:
      tags: [Questionnaires]
      summary: Get active questionnaire template
      description: |
        Returns the single active template and its questions.
        Questions are ordered by position ASC, then created_at ASC.
        If position is null, those items appear last.
      operationId: getActiveQuestionnaire
      responses:
        "200":
          description: Active template with questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActiveQuestionnaireResponse"
        "404":
          description: No active template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/questionnaires:
    post:
      tags: [Questionnaires]
      summary: Create questionnaire template
      description: |
        Creates a new questionnaire template. New templates are created with
        is_active=false by default to enforce a single active template.
      operationId: createQuestionnaire
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionnaireCreateRequest"
      responses:
        "201":
          description: Template created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionnaireCreateResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/v1/questionnaires/{template_id}/questions:
    post:
      tags: [Questionnaires]
      summary: Add questions to a template
      description: |
        Adds one or more questions to the template. If the template is active,
        changes are rejected to preserve instrument stability.
      operationId: addQuestions
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: template_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/QuestionCreateRequest"
                - $ref: "#/components/schemas/QuestionCreateListRequest"
      responses:
        "201":
          description: Questions created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionsCreatedResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/v1/questionnaires/{template_id}/activate:
    post:
      tags: [Questionnaires]
      summary: Activate a template
      description: |
        Activates a template and deactivates all others. Ensures exactly one
        active template exists globally.
      operationId: activateQuestionnaire
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: template_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Template activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/v1/questionnaires/active/clone:
    post:
      tags: [Questionnaires]
      summary: Clone active template
      description: |
        Creates a new draft template by cloning the current active template.
        Requires a new version string. The cloned template is not activated.
      operationId: cloneActiveQuestionnaire
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionnaireCloneRequest"
      responses:
        "201":
          description: Template cloned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionnaireCloneResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token sent in the Authorization header.
    mfaEnrollmentToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived token for MFA setup/confirm only (claim mfa_enrollment=true).
    cookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token
      description: Refresh token cookie (HttpOnly).
    csrfCookie:
      type: apiKey
      in: cookie
      name: csrf_refresh_token
      description: CSRF cookie used for double-submit validation.
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: CSRF header used for double-submit validation.
    metricsToken:
      type: apiKey
      in: header
      name: Authorization
      description: "Bearer token for /metrics (Authorization: Bearer <token>)."
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        msg:
          type: string
        error:
          type: string
      example:
        msg: Invalid credentials
        error: invalid_credentials
...
    MessageResponse:
      type: object
      properties:
        msg:
          type: string
      example:
        msg: ok
...
    QuestionnaireTemplate:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        version:
          type: string
        description:
          type: string
          nullable: true
        is_active:
          type: boolean
    QuestionItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        text:
          type: string
        response_type:
          type: string
        disorder_id:
          type: string
          format: uuid
          nullable: true
        position:
          type: integer
          nullable: true
    ActiveQuestionnaireResponse:
      type: object
      properties:
        questionnaire_template:
          $ref: "#/components/schemas/QuestionnaireTemplate"
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuestionItem"
    QuestionnaireCreateRequest:
      type: object
      required: [name, version]
      properties:
        name:
          type: string
        version:
          type: string
        description:
          type: string
          nullable: true
    QuestionnaireCloneRequest:
      type: object
      required: [version]
      properties:
        version:
          type: string
        name:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
    QuestionnaireCreateResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        version:
          type: string
    QuestionnaireCloneResponse:
      type: object
      properties:
        template_id:
          type: string
          format: uuid
        name:
          type: string
        version:
          type: string
        is_active:
          type: boolean
        question_count:
          type: integer
    QuestionCreateRequest:
      type: object
      required: [code, text, response_type]
      properties:
        code:
          type: string
        text:
          type: string
        response_type:
          type: string
        disorder_id:
          type: string
          format: uuid
          nullable: true
        position:
          type: integer
          nullable: true
    QuestionCreateListRequest:
      type: array
      items:
        $ref: "#/components/schemas/QuestionCreateRequest"
    QuestionsCreatedResponse:
      type: object
      properties:
        template_id:
          type: string
          format: uuid
        created:
          type: integer
...